+++ src/CHIR/Optimization/RangePropagation.cpp
@@ -237,7 +237,7 @@ std::vector<size_t> GetVArraySizeList(const Ptr<Type>& type)
     }
     while (indexType->GetTypeKind() == Type::TypeKind::TYPE_VARRAY) {
         auto vArrayType = StaticCast<const VArrayType*>(indexType);
-        size.push_back(vArrayType->GetSize());
+        size.push_back(static_cast<size_t>(vArrayType->GetSize()));
         indexType = vArrayType->GetElementType();
     }
     return size;
